command/grep: do not write to file if not necessary
Updates #13713.

The workings of :grep/:lgrep/:make/:lmake:

Set defaults:

    :set grepprg=grep\ -n
    :set shellpipe=2>&1\\\|\ tee
    :set makeef=

Execute `:grep`:

    :grep restrictions LICENSE
    !grep -n restrictions LICENSE /dev/null 2>&1| tee /tmp/nvimD2UTQz/1
    [No write since last change]
    LICENSE:214:I)  There are no restrictions on distributing unmodified copies of Vim except
    (1 of 1): I)  There are no restrictions on distributing unmodified copies of Vim except

    <results>

This has the effect of:

 1. Outputting the shell command itself to the command window
 2. Outputting the results to the command window (see `tee`).
 3. Outputting the results to a temporary file (`:help makeef`)

Afterwards, the use must press 'Enter' to continue. They might call
`:copen` to display the quickfix list and see all results (which are
loaded from the temporary file.

This process is cumbersome, spammy and inefficient, which is why people
improve upon it in various ways:

 - `:help grep` has a ":NewGrep" example.
 - https://gist.github.com/romainl/56f0c28ef953ffc157f36cc495947ab3 goes
 - Plugins that are veneers on top of :grep/:lgrep.

This commit changes what it means for `shellpipe` to be unset.
Previously:

> The 'shellpipe' option defaults to ">%s 2>&1" for Win32. This means
> that the output of the compiler is saved in a file and not shown on
> the screen directly.  For Unix "| tee" is used.  The compiler output
> is shown on the screen and saved in a file the same time.  Depending
> on the shell used "|& tee" or "2>&1| tee" is the default, so stderr
> output will be included.
>
> If 'shellpipe' is empty, the {errorfile} part will be omitted.  This
> is useful for compilers that write to an errorfile themselves.

And the previous behaviour:

    :set shellpipe=
    :set makeef=
    :grep restrictions LICENSE
    E40: Can't open errorfile /tmp/nvimD2UTQz/13

After this commit, an empty `shellpipe` and `makeef` means Neovim will
read output directly stdout of `&grepprg/&makeprg`. This is similar to
the common improvement of calling `&grepprg` via `system()`.

The Vim documentation already references that this may go away in the
future, from `:help shellpipe`:

> In the future pipes may be used for filtering and this option will >
become obsolete (at least for Unix).