feat(:grep/:make): do not write to file if not necessary
Update #6117.

Add a new mode of operation for the `:grep/:lgrep/:make/:lmake` Ex
commands: If `&shellpipe` and `&makeef` are unset, read from the stdout
instead of a temporary file. Similar in function to:

    :cgetexpr system(&grepprg ...)

Or

    :silent grep!

Compared to before this commit, grep/make output is no longer writen to
the TTY, and the user no longer has to press enter to return to Neovim.

This mode may obsolete parts of the following:

 - `:help grep` has a ":NewGrep" example.
 - https://gist.github.com/romainl/56f0c28ef953ffc157f36cc495947ab3
 - Plugins that are veneers on top of :grep/:lgrep.
   - https://github.com/george-b/sf.vim

It has the added benefit of working even if the filesystem is fully
read-only (although at that point saving the file you're editing would
become difficult).

Right now, `:grep` has the following behaviour. Defaults (for my shell):

    :set grepprg=grep\ -n
    :set shellpipe=2>&1\\\|\ tee
    :set makeef=

Execute `:grep`:

    :grep restrictions LICENSE
    !grep -n restrictions LICENSE /dev/null 2>&1| tee /tmp/nvimD2UTQz/1
    [No write since last change]
    LICENSE:214:I)  There are no restrictions on distributing unmodified copies of Vim except
    (1 of 1): I)  There are no restrictions on distributing unmodified copies of Vim except

    <results>

This:

  1. Outputting the shell command itself to the command window
  2. Outputting the results to the command window (see `tee`).
  3. Outputting the results to a temporary file (`:help makeef`)

Afterwards, the uses must press 'Enter' to continue. They might call
`:copen` to display the quickfix list and see all results (which are
loaded from the temporary file.

This process is cumbersome, spammy and inefficient, which is why people
improve upon it with scripts.

This commit changes what it means for `shellpipe` to be unset, if
`makeef` is also unset.

Previously:

> The 'shellpipe' option defaults to ">%s 2>&1" for Win32. This means
> that the output of the compiler is saved in a file and not shown on
> the screen directly.  For Unix "| tee" is used.  The compiler output
> is shown on the screen and saved in a file the same time.  Depending
> on the shell used "|& tee" or "2>&1| tee" is the default, so stderr
> output will be included.
>
> If 'shellpipe' is empty, the {errorfile} part will be omitted.  This
> is useful for compilers that write to an errorfile themselves.

Before this commit, unsetting `shellpipe` and `makeef` would result in
the following:

    :set shellpipe=
    :set makeef=
    :grep restrictions LICENSE
    E40: Can't open errorfile /tmp/nvimD2UTQz/13

After this commit, an empty `shellpipe` and `makeef` maeks Neovim behave
smilar to the common improvement of calling `&grepprg` via `system()`.

The Vim documentation already mentions that this may go away in the
future, from `:help shellpipe`:

> In the future pipes may be used for filtering and this option will
> become obsolete (at least for Unix).

One might think that avoiding (temporary) file I/O is a big boon to
performance, and it might be, but it's not on my (Linux) system:

    $ TEST_FILE=test/functional/ex_cmds/grep_spec.lua  make functionaltest
    -------- Global test environment setup.
    -------- Running tests from test/functional/ex_cmds/grep_spec.lua
    RUN      T1 :grep does not hang on large input #2983: 5731.78 ms OK
    RUN      T2 :grep works when both shellpipe and makeef are unset: 5722.86 ms OK
    -------- 2 tests from test/functional/ex_cmds/grep_spec.lua (11498.24 ms total)

Both tests take ~5 seconds to run. The individual grep commands produce
a decent amount of output, but take a negligible time to run:

    $ cd test/ && time grep a **/* >/dev/null
    ...
    grep --color=auto a **/* > /dev/null  0.01s user 0.01s system 96% cpu 0.018 total

I've profiled this and it's due to errorformat interpretation [2]

[1]: I bet it was `errorformat` parsing. I tried to do
     `feed_command([[set errorformat=%f:%l:%m]])` since the default
     errorformat seems very verbose (and each comma-separate pattern is
     a separate regex). That made no difference to the runtime. It may
     still be `errorformat` parsing, but not as straightforward.

[2]:

perf report --percent-limit 10 --call-graph=none | cat - | head -60
    95.24%     0.00%  nvim      [.] __libc_start_main
    95.24%     0.00%  nvim      [.] main
    95.24%     0.00%  nvim      [.] 0x5541d68949564100
    95.22%     0.00%  nvim      [.] normal_enter
    95.22%     0.00%  nvim      [.] state_enter
    95.04%     0.00%  nvim      [.] normal_execute
    89.37%     0.00%  nvim      [.] do_cmdline
    89.37%     0.00%  nvim      [.] nv_colon
    89.36%     0.00%  nvim      [.] do_one_cmd
    89.36%     0.00%  nvim      [.] execute_cmd0
    89.36%     0.00%  nvim      [.] ex_make
    89.22%     0.01%  nvim      [.] qf_init_ext
    89.20%     0.03%  nvim      [.] qf_init_process_nextline
    88.23%     0.12%  nvim      [.] qf_parse_line
    88.01%     0.26%  nvim      [.] qf_parse_get_fields
    87.71%     0.01%  nvim      [.] vim_regexec
    87.69%     0.13%  nvim      [.] vim_regexec_string
    87.57%     0.02%  nvim      [.] nfa_regexec_nl
    87.54%     2.36%  nvim      [.] nfa_regexec_both
    85.16%     0.10%  nvim      [.] nfa_regtry
    82.96%    23.88%  nvim      [.] nfa_regmatch
    55.78%    41.71%  nvim      [.] addstate